+++ src/infra/llm/llm.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { PromptLoader } from 'src/common/utils/prompt.loader';
 import { LLMService } from './llm.service';
+import { RagService } from './rag.service';
 
 @Module({
-    providers: [LLMService, PromptLoader],
-    exports: [LLMService],
+    providers: [LLMService, PromptLoader, RagService],
+    exports: [LLMService, RagService],
 })
 export class LLMModule {}
